Fix wrong cfd rising time correction used in HGTD Digitization
Fix wrong cfd rising time correction

The intended time difference mean value between pulse staring point and the cfd
point is supposed to be 0.408 ns instead of 0.4 ns. And when calculating
the pulse, `int` instead of `size_t` is needed as the sign of `i` matters and
it's desired

+++ HighGranularityTimingDetector/HGTD_Digitization/HGTD_Digitization/src/HGTD_TimingResolution.cxx
@@ -289,7 +289,7 @@ HGTD_TimingResolution::PulseWaveform HGTD_TimingResolution::simulatePulse(
   float p_time;
 
   // Convolution integral of Landau and Gaussian by sum
-  for (size_t i = 412; i < 527; i++) {
+  for (int i = 412; i < 527; i++) {
 
     fland = TMath::Landau(3.5 + (i * 0.005), mpc, pulseParameter[0]) /
             pulseParameter[0];
